% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_to_attributes.R
\name{constant_to_attributes}
\alias{constant_to_attributes}
\title{Constants to attributes}
\usage{
constant_to_attributes(
  x,
  constant_columns,
  attribute_name = "datasetAttributes"
)
}
\arguments{
\item{x}{A data frame or similar R object.}

\item{attribute_name}{The name of the attribute which will containt the \code{constant_columns}.}

\item{constant_colums}{A selection of constant columns. Defaults to \code{'all'}.}
}
\value{
A subsetted data frame, which includes the unique values of the constant columns
among the attributes of the R object.
}
\description{
Move columns that have only a single value present the dataset into the
attributes of a dataset.
}
\examples{
path <- system.file("ZA5688_v6_sample.sav", package = "surveydataset")
survey_raw <- read_spss_declared(path)
survey_raw <- survey_raw[, c(1:5, 7, 10:12 )]
constant_to_attributes(x = survey_raw, constant_columns = 1:5)

my_survey <- constant_to_attributes(x = survey_raw, constant_columns = 1:5)
attr(my_survey, "datasetAttributes")
}
\concept{dimension reduction functions}
