% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spss_declared.R
\name{read_spss_declared}
\alias{read_spss_declared}
\alias{read_sav_declared}
\alias{read_por_declared}
\title{Read and write SPSS files}
\usage{
read_sav_declared(
  file,
  encoding = NULL,
  user_na = FALSE,
  col_select = NULL,
  skip = 0,
  n_max = Inf,
  .name_repair = "unique"
)

read_por_declared(
  file,
  user_na = FALSE,
  col_select = NULL,
  skip = 0,
  n_max = Inf,
  .name_repair = "unique"
)

read_spss_declared(
  file,
  user_na = FALSE,
  col_select = NULL,
  skip = 0,
  n_max = Inf,
  .name_repair = "unique"
)
}
\arguments{
\item{file}{The file to read.}

\item{encoding}{The character encoding used for the file. The default,
\code{NULL}, use the encoding specified in the file, but sometimes this
value is incorrect and it is useful to be able to override it.}

\item{user_na}{If \code{TRUE} variables with user defined missing will
be read into \code{\link[=labelled_spss]{labelled_spss()}} objects. If \code{FALSE}, the
default, user-defined missings will be converted to \code{NA}.}

\item{col_select}{One or more selection expressions, like in
\code{\link[dplyr:select]{dplyr::select()}}. Use \code{c()} or \code{list()} to use more than one expression.
See \code{?dplyr::select} for details on available selection options. Only the
specified columns will be read from \code{data_file}.}

\item{skip}{Number of lines to skip before reading data.}

\item{n_max}{Maximum number of lines to read.}
}
\value{
A tibble, data frame variant with nice defaults.

Variable labels are stored in the "label" attribute of each variable.
It is not printed on the console, but the RStudio viewer will show it.
}
\description{
\code{read_sav()} reads both \code{.sav} and \code{.zsav} files.
\code{read_por_declared()} reads \code{.por} files.
\code{read_spss_declared()} uses either \code{read_por_declared()} or \code{read_sav_declared()} based on the
file extension.
}
\details{
Currently haven can read and write logical, integer, numeric, character
and factors. This is not ideal for social science surveys, and therefore we created a
wrapper haven's functions to read the data into declared types.
}
\examples{
path <- system.file("examples", "ZA5688_v6_sample.sav", package = "surveydataset")
read_sav_declared(path)

tmp <- tempfile(fileext = ".sav")
read_sav_declared(tmp)
}
