% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey}
\alias{survey}
\title{Create a survey}
\usage{
survey(
  x,
  Dimensions,
  Measures,
  Attributes,
  Constants,
  Title = NULL,
  Identifier,
  Label = NULL,
  Creator = NULL,
  Publisher = NULL,
  Issued = NULL,
  Subject = NULL
)
}
\arguments{
\item{x}{A data.frame or inherited tibble, data.frame, or a structured list.}

\item{Dimensions}{The name or column number of the dimensions within the dataset.}

\item{Measures}{The name or column number of the measures within the dataset.}

\item{Attributes}{The name or column number of the attributes within the dataset.}

\item{Title}{\href{https://purl.org/dc/elements/1.1/title}{dct:title}, a name given to the resource.
\code{\link[dataset]{datacite}} allows the use of alternate titles, too. See \code{\link[dataset]{dataset_title}}.}

\item{Identifier}{An unambiguous reference to the resource within a given context.
Recommended practice is to identify the resource by means of a string conforming to an
identification system. Examples include International Standard Book Number (ISBN),
Digital Object Identifier (DOI), and Uniform Resource Name (URN).
Select and identifier scheme from
\href{https://www.ukoln.ac.uk/metadata/dcmi-ieee/identifiers/index.html}{registered URI schemes maintained by IANA}.
More details: \href{https://www.ukoln.ac.uk/metadata/dcmi-ieee/identifiers/}{Guidelines for using resource identifiers in Dublin Core metadata and IEEE LOM}.
Similar to \code{Identifier} in \code{\link[dataset]{datacite}}. See \code{\link[dataset]{identifier}}.}

\item{Label}{may be used to provide a human-readable version of the dataset's name.
A text description (optionally with a language tag) as defined by
\href{https://www.w3.org/TR/rdf-schema/#ch_label}{rdfs:label}.}

\item{Creator}{An entity primarily responsible for making the resource. \href{https://purl.org/dc/elements/1.1/creator}{dct:creator}
Corresponds to \code{Creator} in \code{\link[dataset]{datacite}}. See \code{\link[dataset]{creator}}.}

\item{Publisher}{Corresponds to \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#publisher}{dct:publisher}
and Publisher in DataCite.
The name of the entity that holds, archives, publishes prints, distributes, releases,
issues, or produces the resource. This property will be used to formulate the citation,
so consider the prominence of the role. For software, use \code{Publisher} for the
code repository. If there is an entity other than a code repository, that
"holds, archives, publishes, prints, distributes, releases, issues, or produces" the
code, use the property Contributor/contributorType/hostingInstitution for the code
repository. See \code{\link[dataset]{publisher}}.}

\item{Issued}{Corresponds to \href{https://purl.org/dc/elements/1.1/date}{dct:date}.}

\item{Subject}{Recommended for discovery in DataCite. Subject, keyword, classification code, or key
phrase describing the resource. Similar to \href{https://purl.org/dc/elements/1.1/subject}{dct:subject}. \cr
Use \code{\link[dataset]{subject}} to properly add a key phrase from a controlled vocabulary
and create structured Subject objects with \code{\link[dataset]{subject_create}}.}
}
\value{
A well-described and formatted data frame with survey data and rich metadata.
}
\description{
Create a dataset suitable for working with social sciences surveyd data.
}
\examples{
path <- system.file("ZA5688_v6_sample.sav", package = "surveydataset")
survey_raw       <- read_spss_declared(path)

Identifier <- survey_raw$doi[1]

y <- survey(
  x = survey_raw,
  Dimensions = c(10, 22:33),
  Measures = c(13:21),
  Attributes = c(7:9, 11),
  Constants = c(1:6),
  Title = "Test Survey",
  Identifier = survey_raw$doi[1],
  Label = as.character(as.factor(survey_raw$survey[1]))
)
}
